requirements.txt
Django>=5.1.1
djangorestframework
psycopg2-binary
redis
django-redis
pytest
pytest-django
factory_boy
pytest-cov
drf-yasg
